% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_surv_ranger_cox.R
\name{LearnerSurvRangerCox}
\alias{LearnerSurvRangerCox}
\title{R6 Class to construct a Ranger survival learner for Cox regression}
\description{
R6 Class to construct a Ranger survival learner for Cox regression

R6 Class to construct a Ranger survival learner for Cox regression
}
\examples{

## ------------------------------------------------
## Method `LearnerSurvRangerCox$new`
## ------------------------------------------------

LearnerSurvRangerCox$new()

}
\section{Super class}{
\code{\link[mlexperiments:MLLearnerBase]{mlexperiments::MLLearnerBase}} -> \code{LearnerSurvRangerCox}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvRangerCox-new}{\code{LearnerSurvRangerCox$new()}}
\item \href{#method-LearnerSurvRangerCox-clone}{\code{LearnerSurvRangerCox$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="bayesian_scoring_function"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-bayesian_scoring_function'><code>mlexperiments::MLLearnerBase$bayesian_scoring_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="cross_validation"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-cross_validation'><code>mlexperiments::MLLearnerBase$cross_validation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="fit"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-fit'><code>mlexperiments::MLLearnerBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="predict"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-predict'><code>mlexperiments::MLLearnerBase$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvRangerCox-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvRangerCox-new}{}}}
\subsection{Method \code{new()}}{
Create a new `LearnerSurvRangerCox` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvRangerCox$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `LearnerSurvRangerCox` R6 object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{LearnerSurvRangerCox$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvRangerCox-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvRangerCox-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvRangerCox$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
